jacobian is working
